+++ examples/color.rs
@@ -31,7 +31,7 @@ fn main() -> Result<(), TBLError<String>> {
     .with_length(90)
     .with_renderer(&render)
     .render()?;
-    for line in rendered {
+    for line in rendered.iter().flatten() {
         println!("{}", line);
     }
     Ok(())